Trait janus_aggregator::aggregator::query_type::CollectableQueryType

pub trait CollectableQueryType: CoreCollectableQueryType + AccumulableQueryType {
    // Required method
    fn validate_query_count<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait, const SEED_SIZE: usize, C, A>(
        tx: &'life0 Transaction<'life1, C>,
        vdaf: &'life2 A,
        task: &'life3 AggregatorTask,
        batch_identifier: &'life4 Self::BatchIdentifier,
        aggregation_param: &'life5 A::AggregationParam
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where A::AggregationParam: Send + Sync + Eq + Hash,
             C: 'async_trait + Clock,
             A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             'life4: 'async_trait,
             'life5: 'async_trait;
}

CollectableQueryType represents a query type that can be collected by Janus. This trait extends AccumulableQueryType with additional functionality required for collection.

Required Methods

fn validate_query_count<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait, const SEED_SIZE: usize, C, A>( tx: &'life0 Transaction<'life1, C>, vdaf: &'life2 A, task: &'life3 AggregatorTask, batch_identifier: &'life4 Self::BatchIdentifier, aggregation_param: &'life5 A::AggregationParam ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where A::AggregationParam: Send + Sync + Eq + Hash, C: 'async_trait + Clock, A: 'async_trait + Aggregator<SEED_SIZE, 16> + Send + Sync, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, 'life4: 'async_trait, 'life5: 'async_trait,

Validates query count for a given batch, per the size checks in https://www.ietf.org/archive/id/draft-ietf-ppm-dap-02.html#section-4.5.6.