Module janus_aggregator::binary_utils

Utilities for Janus binaries.

Modules

job_driver

Discovery and driving of jobs scheduled elsewhere.

Structs

BinaryContext

BinaryContext provides contextual objects related to a Janus binary.

CommonBinaryOptions

Common options that are used by all Janus binaries.

Traits

BinaryOptions

Options for Janus binaries.

Functions

database_pool

Connects to a database, given a config. db_password is mutually exclusive with the database password specified in the connection URL in db_config.

datastore

Connects to a datastore, given a connection pool to the underlying database. datastore_keys is a list of AES-128-GCM keys, encoded in base64 with no padding, used to protect secret values stored in the datastore; it must not be empty.

janus_main

read_config

Reads, parses, and returns the config referenced by the given options, or None if no config file path was set.

setup_server

Construct a server that listens on the provided SocketAddr and services requests with handler. If the SocketAddr’s port is 0, an ephemeral port is used. Returns a SocketAddr representing the address and port the server are listening on and a future that can be awaited to wait until the server shuts down.

setup_signal_handler

Register a signal handler for SIGTERM, and return a future that will become ready when a SIGTERM signal is received.