Function janus_aggregator::binary_utils::datastore

pub async fn datastore<C: Clock>(
    pool: Pool,
    clock: C,
    datastore_keys: &[String],
    check_schema_version: bool
) -> Result<Datastore<C>>

Connects to a datastore, given a connection pool to the underlying database. datastore_keys is a list of AES-128-GCM keys, encoded in base64 with no padding, used to protect secret values stored in the datastore; it must not be empty.