Struct janus_aggregator::cache::PeerAggregatorCache

pub struct PeerAggregatorCache { /* private fields */ }

Caches taskprov PeerAggregator’s. This cache is never invalidated, so the process needs to be restarted if there are any changes to peer aggregators.

Implementations

impl PeerAggregatorCache

pub async fn new<C: Clock>(datastore: &Datastore<C>) -> Result<Self, Error>

pub fn get(&self, endpoint: &Url, role: &Role) -> Option<&PeerAggregator>